(TK-38) - (feature) - Refactor tests to use ViewSetTestMixin and consolidate user-related fixtures
