+++ gardeniq/users/tests/tests_views/test_users.py
@@ -6,133 +6,16 @@ Uses the GIVEN-WHEN-THEN pattern for clarity.
 """
 import pytest
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Group, Permission
-from django.contrib.contenttypes.models import ContentType
 from rest_framework import status
 from rest_framework.reverse import reverse
-from rest_framework.test import APIClient
+
+from gardeniq.base.utils.tests import ViewSetTestMixin
 
 User = get_user_model()
 
 
-@pytest.fixture
-def api_client():
-    """Return a DRF API client."""
-    return APIClient()
-
-
-@pytest.fixture
-def user_content_type():
-    """Return the content type for User model."""
-    return ContentType.objects.get_for_model(User)
-
-
-@pytest.fixture
-def test_permission(user_content_type):
-    """Create a test permission."""
-    return Permission.objects.create(
-        codename='test_permission',
-        name='Test Permission',
-        content_type=user_content_type
-    )
-
-
-@pytest.fixture
-def test_permission_2(user_content_type):
-    """Create a second test permission."""
-    return Permission.objects.create(
-        codename='test_permission_2',
-        name='Test Permission 2',
-        content_type=user_content_type
-    )
-
-
-@pytest.fixture
-def test_group(test_permission):
-    """Create a test group with permissions."""
-    group = Group.objects.create(name='Test Group')
-    group.permissions.add(test_permission)
-    return group
-
-
-@pytest.fixture
-def test_group_2(test_permission_2):
-    """Create a second test group with permissions."""
-    group = Group.objects.create(name='Test Group 2')
-    group.permissions.add(test_permission_2)
-    return group
-
-
-@pytest.fixture
-def regular_user():
-    """Create a regular non-staff user."""
-    return User.objects.create_user(
-        username='regularuser',
-        email='regular@example.com',
-        password='RegularPass123!',
-        first_name='Regular',
-        last_name='User'
-    )
-
-
-@pytest.fixture
-def admin_user():
-    """Create an admin user with staff privileges."""
-    return User.objects.create_user(
-        username='adminuser',
-        email='admin@example.com',
-        password='AdminPass123!',
-        first_name='Admin',
-        last_name='User',
-        is_staff=True,
-        is_superuser=True
-    )
-
-
-@pytest.fixture
-def regular_user_with_groups(regular_user, test_group, test_permission):
-    """Create a regular user with groups and permissions."""
-    regular_user.groups.add(test_group)
-    regular_user.user_permissions.add(test_permission)
-    return regular_user
-
-
-@pytest.fixture
-def authenticated_client(api_client, regular_user):
-    """Return an authenticated API client for a regular user."""
-    api_client.force_authenticate(user=regular_user)
-    return api_client
-
-
-@pytest.fixture
-def admin_client(api_client, admin_user):
-    """Return an authenticated API client for an admin user."""
-    api_client.force_authenticate(user=admin_user)
-    return api_client
-
-
-@pytest.fixture
-def unauthenticated_client(api_client):
-    """Return an unauthenticated API client."""
-    api_client.force_authenticate(user=None)
-    return api_client
-
-
-@pytest.fixture
-def user_payload():
-    """Return a valid user creation payload."""
-    return {
-        'username': 'newuser',
-        'password': 'NewUserPass123!',
-        'password_confirm': 'NewUserPass123!',
-        'email': 'newuser@example.com',
-        'first_name': 'New',
-        'last_name': 'User'
-    }
-
-
 @pytest.mark.django_db
-class TestUserAPIModelViewList:
+class TestUserAPIModelViewList(ViewSetTestMixin):
     """Tests for listing users (GET /api/users/)."""
 
     def test_list_users_as_admin(self, admin_client, regular_user, admin_user):
@@ -177,7 +60,7 @@ class TestUserAPIModelViewList:
 
 
 @pytest.mark.django_db
-class TestUserAPIModelViewRetrieve:
+class TestUserAPIModelViewRetrieve(ViewSetTestMixin):
     """Tests for retrieving user details (GET /api/users/{id}/)."""
 
     def test_retrieve_own_user_as_regular_user(self, authenticated_client, regular_user):
@@ -251,7 +134,7 @@ class TestUserAPIModelViewRetrieve:
 
 
 @pytest.mark.django_db
-class TestUserAPIModelViewCreate:
+class TestUserAPIModelViewCreate(ViewSetTestMixin):
     """Tests for creating users (POST /api/users/)."""
 
     def test_create_user_as_admin(self, admin_client, user_payload):
@@ -424,7 +307,7 @@ class TestUserAPIModelViewCreate:
 
 
 @pytest.mark.django_db
-class TestUserAPIModelViewUpdate:
+class TestUserAPIModelViewUpdate(ViewSetTestMixin):
     """Tests for updating users (PUT /api/users/{id}/)."""
 
     def test_update_own_user_as_regular_user(self, authenticated_client, regular_user):
@@ -569,7 +452,7 @@ class TestUserAPIModelViewUpdate:
 
 
 @pytest.mark.django_db
-class TestUserAPIModelViewDelete:
+class TestUserAPIModelViewDelete(ViewSetTestMixin):
     """Tests for deleting users (DELETE /api/users/{id}/)."""
 
     def test_delete_user_as_admin(self, admin_client, regular_user):
@@ -630,7 +513,7 @@ class TestUserAPIModelViewDelete:
 
 
 @pytest.mark.django_db
-class TestUserAPIModelViewPermissions:
+class TestUserAPIModelViewPermissions(ViewSetTestMixin):
     """Tests for permission edge cases and special scenarios."""
 
     def test_regular_user_cannot_assign_groups(
